feat: move storage form-only context into /storage folder
