+++ frontend/src/features/admin-form/responses/ResponsesPage/email/index.ts
@@ -1,0 +1,1 @@
+export { EmailResponsesTab } from './EmailResponsesTab'
