+++ frontend/src/features/admin-form/responses/ResponsesPage/storage/index.ts
@@ -1,0 +1,1 @@
+export { StorageResponsesTab } from './StorageResponsesTab'
